Added moving platforms
The velocity, acceleration and turnaround delay are all tunable
parameters. The platform can start at an arbitrary offset into its cycle
also.

Created a scene where the player is riding a vertical moving platform.

NOTE: that if the player is on the platform and the platform is moving
horizontally then the player will fall off if he doesn't match the x
velocity